upd: lessons

+++ 1/bitwise_operators.cpp
@@ -17,7 +17,7 @@ int main(int argc, char** argv)
     std::cout << (4 << 2) << std::endl;
     std::cout << (4 >> 2) << std::endl;
     std::cout << (~int(5)) << std::endl;
-    std::cout << (~u_int64_t(5)) << std::endl;
+    std::cout << (~u_int32_t(5)) << std::endl;
     std::cout << (12 ^ 10) << std::endl;
     return 0;
 }
